services/data-collection/requirements.txt
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
dnspython==2.6.1
filelock==3.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
joblib==1.3.2
langcodes==3.3.0
langdetect==1.0.9
lxml==5.1.0
nltk==3.8.1
packaging==23.2
parsel==1.8.1
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==23.12.11
pycparser==2.21
PyDispatcher==2.0.7
pymongo==4.6.2
pyOpenSSL==24.0.0
python-dotenv==1.0.1
queuelib==1.6.2
redis==5.0.2
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
schedule==1.2.1
Scrapy==2.11.1
Scrapy-UserAgents==0.0.1
service-identity==24.1.0
six==1.16.0
tldextract==5.1.1
tqdm==4.66.2
Twisted==24.3.0
typing_extensions==4.10.0
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.2
